refactor(server): extract workspace lifecycle routes
